Data from scripts 1 and 2 for running scripts 3 and 4.
These contain the data produced in script 2 (script 1 is all but object RLS_19 used in script 2 is included). Raw RLS data are available from www.reeflifesurvey.com, we provide data outputs after pre-processing in script 1. The main output from script 2 is an object called ThermalNicheData_New which is used in modelling for script 3, and data transformations in script 4 which therefore provides the data and code to reproduce the key results in the main manuscript.